Timezone correction in departure_alarm.py
This pdf briefly describes the issue with the current code in the Udemy course which would result in students in different timezones getting different results and training their models on incorrect trip data. A simple solution is proposed (adding 3 lines of code) to fix this by setting the timezone for the current Python instance to US/Eastern so that nothing else in the code needs to be changed in order to pull the correct trip data from the MBTA database.