Fix to account for different timezones
This smaller version of the Udemy course code just pulls trip data for the dates between May 1, 2018 and May 5, 2018 (inclusive), and displays the number of trips that lie between 7 AM and 10 AM on each day. Small modifications have been made to display the current timezone in your Python instance, and also some descriptive print statements about what is being displayed. Note that this version of the program will give you the desired results even if your system timezone is not US/Eastern, as we now explicitly set our timezone within the Python instance to be US/Eastern